requirements.txt
jira>=1.0.10
google-api-python-client
python-dateutil>=2.6.1
PyYAML>=3.12
